Refactor component ID handling to use size_t and centralize logic
Replaces manual integer IDs with size_t for component identification and centralizes ID assignment in the base component class. Updates all usages and event structures to use the new ID type, removes redundant ID logic from button and scene classes, and improves consistency across component registration, lookup, and event posting. Also updates includes to use <cstddef> and removes unused spdlog/fmt/fmt.h include.

+++ src/engine/scenes/game_overlay.hpp
@@ -7,6 +7,7 @@
 #include "../result.hpp"
 #include "scene.hpp"
 
+#include <cstddef>
 #include <string>
 
 namespace engine {
@@ -21,7 +22,7 @@ public:
 	auto layout(size screen_size) -> result<> override;
 
 private:
-	int version_display_ = 0;
+	size_t version_display_ = 0;
 	static constexpr auto margin = 10.0F;
 	int click_{0};
 	static auto open_url(const std::string &url) -> result<>;
